♻️ Refactor: Date Extension 정의 (#1)
mangers 제거
